# src/main/resources/yapi_export_api/swaggerApi.json
{
  "swagger": "2.0",
  "info": {
    "title": "ERP-METRICS",
    "version": "last",
    "description": "erp-metrics 雷神架构组项目"
  },
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/app/monitor/data/list": {
      "post": {
        "tags": [
          "应用异常监控"
        ],
        "summary": "应用监控数据列表",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "root",
            "in": "body",
            "schema": {
              "type": "object",
              "title": "empty object",
              "properties": {
                "configId": {
                  "type": "integer",
                  "description": "应用配置的id"
                },
                "startTime": {
                  "type": "string",
                  "description": "起始时间"
                },
                "endTime": {
                  "type": "string",
                  "description": "起止时间"
                },
                "pageNum": {
                  "type": "integer",
                  "mock": {
                    "mock": "1"
                  },
                  "description": "第几页"
                },
                "pageSize": {
                  "type": "string",
                  "mock": {
                    "mock": "10"
                  },
                  "description": "每页大小"
                },
                "appOwner": {
                  "type": "string",
                  "mock": {
                    "mock": "冯立彬"
                  },
                  "description": "应用配置负责人"
                },
                "deptId": {
                  "type": "number",
                  "mock": {
                    "mock": "1"
                  },
                  "description": "应用归属项目部ID"
                }
              },
              "required": [
                "pageNum",
                "pageSize"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n    \"msg\": \"成功\",\n    \"result\": 1,\n    \"data\": {\n        \"records\": [\n            {\n                \"id\": \"1640\",\n                \"appName\": \"ERP应用空指针异常监控\",\n                \"appOwner\": \"冯立彬\",\n                \"errCount\": 4,\n                \"kibanaUrl\": \"http://127.0.0.1:5601/app/kibana#/discover?_a=(columns:!(_source),filters:!(('$state':(store:appState),meta:(alias:!n,disabled:!f,index:'44104a00-b46b-11ea-8e34-23aecccf7569',key:query,negate:!f,type:custom,value:'{\\\"bool\\\":{\\\"must_not\\\":[{\\\"term\\\":{\\\"fields.service.keyword\\\":{\\\"value\\\":\\\"ld-erp-devops\\\"}}}],\\\"must\\\":[{\\\"wildcard\\\":{\\\"message\\\":{\\\"value\\\":\\\"*nullpointerexception*\\\"}}}]}}'),query:(bool:(must_not:!((term:(fields.service.keyword:(value:ld-erp-devops)))),must:!((wildcard:(message:(value:'*nullpointerexception*')))))))),index:'44104a00-b46b-11ea-8e34-23aecccf7569',interval:auto,query:(language:kuery,query:''),sort:!('@timestamp',desc))&_g=(filters:!(),refreshInterval:(pause:!t,value:0),time:(from:'2021-01-14T15:21:00',to:'2021-01-14T15:31:00'))\",\n                \"remark\": null,\n                \"status\": 0,\n                \"closeTime\": null,\n                \"startTime\": \"2021-01-14 15:21:00\",\n                \"endTime\": \"2021-01-14 15:31:00\",\n\t\t\"deptId\":1,\n\t\t\"deptName\":\"雷神项目部\"\n            },\n            {\n                \"id\": \"1616\",\n                \"appName\": \"ERP表情符号异常监控\",\n                \"appOwner\": \"冯立彬\",\n                \"errCount\": 2,\n                \"kibanaUrl\": \"http://127.0.0.1:5601/app/kibana#/discover?_a=(columns:!(_source),filters:!(('$state':(store:appState),meta:(alias:!n,disabled:!f,index:'44104a00-b46b-11ea-8e34-23aecccf7569',key:query,negate:!f,type:custom,value:'{\\\"bool\\\":{\\\"must\\\":[{\\\"wildcard\\\":{\\\"message\\\":{\\\"value\\\":\\\"*incorrect*\\\"}}},{\\\"wildcard\\\":{\\\"message\\\":{\\\"value\\\":\\\"*string*\\\"}}}]}}'),query:(bool:(must:!((wildcard:(message:(value:'*incorrect*'))),(wildcard:(message:(value:'*string*')))))))),index:'44104a00-b46b-11ea-8e34-23aecccf7569',interval:auto,query:(language:kuery,query:''),sort:!('@timestamp',desc))&_g=(filters:!(),refreshInterval:(pause:!t,value:0),time:(from:'2021-01-14T12:02:00',to:'2021-01-14T12:12:00'))\",\n                \"remark\": null,\n                \"status\": 0,\n                \"closeTime\": null,\n                \"startTime\": \"2021-01-14 12:02:00\",\n                \"endTime\": \"2021-01-14 12:12:00\",\n\t\t\"deptId\":1,\n\t\t\"deptName\":\"雷神项目部\"\n            }\n        ],\n        \"total\": 70,\n        \"size\": 25,\n        \"current\": 1,\n        \"searchCount\": true,\n        \"pages\": 3\n    }\n}"
            }
          }
        }
      }
    },
...
    "/app/monitor/data/updateRemark": {
      "post": {
        "tags": [
          "应用异常监控"
        ],
        "summary": "更新告警任务备注",
        "description": "",
        "consumes": [
          "text/plain"
        ],
        "parameters": [
          {
            "name": "raw",
            "in": "body",
            "description": "raw paramter",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n    \"id\": \"350\",\n    \"remark\": \"这是一个备注\"\n}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n    \"msg\": \"成功\",\n    \"result\": 1,\n    \"data\": true\n}"
            }
          }
        }
      }
    },
    "/app/monitor/data/updateStatus": {
      "post": {
        "tags": [
          "应用异常监控"
        ],
        "summary": "更新告警任务的状态",
        "description": "",
        "consumes": [
          "text/plain"
        ],
        "parameters": [
          {
            "name": "raw",
            "in": "body",
            "description": "raw paramter",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n    \"id\": \"260\",\n    \"status\": \"1\",\n    \"access_token\": \"5axdlkh1i3nr4xydtfzrjuhx61s5eo6\"\n}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string",
              "format": "binary",
              "default": "{\n    \"msg\": \"成功\",\n    \"result\": 1,\n    \"data\": true\n}"
            }
          }
        }
      }
    },
    "/app/monitor/data/detail": {
      "get": {
        "tags": [
          "应用异常监控"
        ],
        "summary": "获取告警任务详情",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "告警任务的ID",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "title": "empty object",
              "properties": {}
            }
          }
        }
      }
    },
...
  }
}